+++ src/components/layout/list/List.tsx
@@ -1,11 +1,18 @@
 import React from "react";
+import { Header } from "./Header";
 
 import styles from "./List.module.scss";
 
-const List: React.FC = (props) => {
+interface SubComponents {
+    Header: typeof Header;
+}
+
+const List: React.FC & SubComponents = (props) => {
     return (
         <ul className={styles.list}>{props.children}</ul>
     );
 };
+
+List.Header = Header;
 
 export { List };