+++ src/index.ts
@@ -1,11 +1,15 @@
 export * from "./components/form/Checkbox.js";
+export * from "./components/form/Toggle.js";
 
 export * from "./components/icons/Icon.js";
 
 export * from "./components/layout/header/Header.js";
+export * from "./components/layout/header/HeaderIconAction.js";
 export * from "./components/layout/header/StickyHeader.js";
 export * from "./components/layout/list/List.js";
 export * from "./components/layout/list/Item.js";
 
+export * from "./components/ui/action/Action.js";
+export * from "./components/ui/action/EqualActions.js";
 export * from "./components/ui/button/Button.js";
 export * from "./components/ui/directionPad/Pad.js";